init project test

+++ src/helpers/delete_images.js
@@ -1,10 +1,13 @@
+/* eslint-disable prefer-promise-reject-errors */
+/* eslint-disable no-unused-expressions */
 const { unlink } = require('fs');
 
-const deleteImages = (path) => {
-  unlink(path, (error) => {
-    if (error) console.log(error);
-    return true;
-  });
-};
+const deleteImages = (path) => new Promise((resolve) => {
+  if (path !== 'public/images/blank.jpg') {
+    unlink(path, (error) => {
+      (error) ? resolve(false) : resolve(true);
+    });
+  }
+});
 
 module.exports = deleteImages;
